Decided to keep the structure and fetch in the MessageBoard
