+++ code/src/App.js
@@ -1,24 +1,24 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 
 import MessageBoard from './components/MessageBoard'
 import MessageForm from './components/MessageForm'
 
-import { API_URL } from './reusable/urls'
+//import { API_URL } from './reusable/urls'
 
 export const App = () => {
   const [thoughts, setThoughts] = useState([])
-    useEffect(() => {
-    fetchThoughts()
-  }, [thoughts]);
+  //   useEffect(() => {
+  //   fetchThoughts()
+  // }, [thoughts]);
 
-  const fetchThoughts = () => {
-    fetch(API_URL)
-      .then(response => response.json())
-      .then(data => {
-        console.log(data)
-        setThoughts(data)
-      })
-    }
+  // const fetchThoughts = () => {
+  //   fetch(API_URL)
+  //     .then(response => response.json())
+  //     .then(data => {
+  //       //console.log(data)
+  //       setThoughts(data)
+  //     })
+  //   }
   return (
     <div>
       <MessageForm
